fixed eslint errors in all files
