+++ visuol/.eslintrc.js
@@ -3,6 +3,7 @@ module.exports = {
     browser: true,
     es2021: true,
   },
+  parser: 'babel-eslint',
   extends: [
     'plugin:react/recommended',
     'airbnb',
@@ -18,5 +19,7 @@ module.exports = {
     'react',
   ],
   rules: {
+    'max-classes-per-file': 'off',
+    'react/prop-types': 'off',
   },
 };
