+++ visuol/src/App.jsx
@@ -1,0 +1,29 @@
+// Css is used from https://ant.design/components/layout/
+import './App.css';
+import 'antd/dist/antd.css';
+import { BrowserRouter, useHistory } from 'react-router-dom';
+import { Layout } from 'antd';
+import SideMenu from './Menu';
+import Routes from './Routes';
+
+const { Sider, Content } = Layout;
+const App = () => {
+  const history = useHistory();
+
+  return (
+    <BrowserRouter history={history}>
+      <Layout>
+        <Layout style={{ height: '100vh' }}>
+          <Sider style={{ backgroundColor: '#C1DE9C' }} width={256} className="site-layout-background">
+            <SideMenu loggedIn={'token' in localStorage && localStorage.getItem('expiration') > Date.now() / 1000} />
+          </Sider>
+          <Content style={{ padding: '48px 24px 24px' }}>
+            <Routes loggedIn={'token' in localStorage && localStorage.getItem('expiration') > Date.now() / 1000} />
+          </Content>
+        </Layout>
+      </Layout>
+    </BrowserRouter>
+  );
+};
+
+export default App;
