--- visuol/src/Home.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-class Home extends Component {
-  render() {
-    return ( 
-      <div>
-          <h1>Home Page (Fill this in later)</h1>
-      </div>
-    );
-  }
-};
-
-export default Home;
