+++ visuol/src/Home.jsx
@@ -1,0 +1,13 @@
+import React, { PureComponent } from 'react';
+
+class Home extends PureComponent {
+  render() {
+    return (
+      <div>
+        <h1>Home Page (Fill this in later)</h1>
+      </div>
+    );
+  }
+}
+
+export default Home;
