--- visuol/src/MockData.js
@@ -1,7 +1,0 @@
-export const mock = {
-    company : "Snapchat",
-    role : "Level 3 SWE",
-    base : 105000,
-    bonus : 60000,
-    rsu : 40000
-}
