+++ visuol/src/Registration.jsx
@@ -1,40 +1,45 @@
-//Credit to https://ant.design/components/form/
-//Component for logging in
-import { Form, Input, Button, Switch} from 'antd';
+// Credit to https://ant.design/components/form/
+// Component for logging in
+import {
+  Form, Input, Button, Switch,
+} from 'antd';
 import React, { Component } from 'react';
 import 'antd/dist/antd.css';
 import axios from 'axios';
 import { withRouter } from 'react-router';
+
 class Registration extends Component {
   constructor(props) {
-    super(props)
+    super(props);
 
     this.state = {
-      displayCompanyForm : false,
-    }
+      displayCompanyForm: false,
+    };
   }
 
-  //OnFinish dictates what happens when submitting a form object
-  finishRegistration = (response) => {
+  // OnFinish dictates what happens when submitting a form object
+  finishRegistration = () => { // (response) => {
+    const { history } = this.props;
     console.log('Success!');
-    this.props.history.push("/login"); //TODO: better convey that registration succeeded (alert?)
+    history.push('/login'); // TODO: better convey that registration succeeded (alert?)
   }
 
-  //OnFinish dictates what happens when submitting a form object
+  // OnFinish dictates what happens when submitting a form object
   onFinish = (values) => {
+    const { displayCompanyForm } = this.state;
+
     const info = {
       name: values.name,
       email: values.email,
       password: values.password,
-      company: (this.state.displayCompanyForm) ? values.company : null,
-    }
-    axios.post('http://localhost:5000/api_v1/register_user', info).then((response) =>
-      {
-        console.log(response);
-        this.finishRegistration(response);
-      }).catch(function(error){
-        console.log(error);
-      });
+      company: (displayCompanyForm) ? values.company : null,
+    };
+    axios.post('http://localhost:5000/api_v1/register_user', info).then((response) => {
+      console.log(response);
+      this.finishRegistration(response);
+    }).catch((error) => {
+      console.log(error);
+    });
   };
 
   onFinishFailed = (errorInfo) => {
@@ -42,13 +47,13 @@ class Registration extends Component {
   };
 
   onClickSwitch = () => {
-    this.setState((state) => {
-      return {displayCompanyForm: !state.displayCompanyForm};
-    });
+    this.setState((state) => ({ displayCompanyForm: !state.displayCompanyForm }));
   };
 
   render() {
-    return ( 
+    const { displayCompanyForm } = this.state;
+
+    return (
       <Form
         name="basic"
         labelCol={{
@@ -103,11 +108,15 @@ class Registration extends Component {
           <Input.Password />
         </Form.Item>
 
-        <Form.Item wrapperCol={{
+        <Form.Item
+          wrapperCol={{
             offset: 4,
             span: 2,
-          }} label="Company Account" valuePropName="checked">
-          <Switch onClick={this.onClickSwitch}/>
+          }}
+          label="Company Account"
+          valuePropName="checked"
+        >
+          <Switch onClick={this.onClickSwitch} />
         </Form.Item>
 
         <Form.Item
@@ -115,12 +124,12 @@ class Registration extends Component {
           name="company"
           rules={[
             {
-              required: this.state.displayCompanyForm,
+              required: displayCompanyForm,
               message: 'Please input your company name',
             },
           ]}
         >
-          <Input disabled={!this.state.displayCompanyForm} />
+          <Input disabled={!displayCompanyForm} />
         </Form.Item>
 
         <Form.Item
@@ -129,13 +138,13 @@ class Registration extends Component {
             span: 10,
           }}
         >
-          <Button type="primary" htmlType="submit" style={{ background: "grey", borderColor: "grey" }}>
+          <Button type="primary" htmlType="submit" style={{ background: 'grey', borderColor: 'grey' }}>
             Submit
           </Button>
         </Form.Item>
       </Form>
     );
   }
-};
+}
 
 export default withRouter(Registration);
