+++ visuol/src/examples/MockData.js
@@ -1,0 +1,7 @@
+export default {
+  company: 'Snapchat',
+  role: 'Level 3 SWE',
+  base: 105000,
+  bonus: 60000,
+  rsu: 40000,
+};
